(GH-3) Add Appveyor CI test
This commit updates build.ps1 for running automated unit tests within Appveyor.
This commit also disables the use of PSDepends as it was casuing issues with
AppVeyor not importing the PoshBot module due to function naming conflicts with
the Configuration module.  This commit also modidies the behavior of Pester to
output the test results when running in an AppVeyor build environment.